+++ script/barrelize.js
@@ -34,6 +34,15 @@ const ignoredTypes = [
   'SignalingProtocol',
   'index',
   'ScreenSignalingProtocol',
+
+  // Events ingestion buffer's JSON interfaces.
+  'JSONIngestionPayloadItem',
+  'JSONIngestionBufferItem',
+  'JSONIngestionEvent',
+  'JSONIngestionRecord',
+
+  // Events ingestion internal functions.
+  'flattenEventAttributes'
 ];
 
 walk('src')